+++ app/javascript/dashboard/routes/dashboard/helpcenter/components/PortalListItem.vue
@@ -6,7 +6,9 @@
         <header>
           <div>
             <div class="title-status--wrap">
-              <h2 class="portal-title block-title">{{ portal.name }}</h2>
+              <h2 class="portal-title block-title">
+                {{ portal.name }}
+              </h2>
               <Label
                 :title="status"
                 :color-scheme="labelColor"
@@ -68,7 +70,7 @@
               size="small"
               icon="delete"
               class="header-action-buttons"
-              @click="onClickDelete(portal)"
+              @click="onClickOpenDeleteModal(portal)"
             />
           </div>
         </header>
@@ -88,7 +90,9 @@
                     'HELP_CENTER.PORTAL.PORTAL_SETTINGS.LIST_ITEM.PORTAL_CONFIG.ITEMS.NAME'
                   )
                 }}</label>
-                <span class="text-block-title">{{ portal.name }}</span>
+                <span class="text-block-title">
+                  {{ portal.name }}
+                </span>
               </div>
               <div class="configuration-item">
                 <label>{{
@@ -96,7 +100,9 @@
                     'HELP_CENTER.PORTAL.PORTAL_SETTINGS.LIST_ITEM.PORTAL_CONFIG.ITEMS.DOMAIN'
                   )
                 }}</label>
-                <span class="text-block-title">{{ portal.custom_domain }}</span>
+                <span class="text-block-title">
+                  {{ portal.custom_domain }}
+                </span>
               </div>
             </div>
             <div class="configuration-items">
@@ -106,7 +112,9 @@
                     'HELP_CENTER.PORTAL.PORTAL_SETTINGS.LIST_ITEM.PORTAL_CONFIG.ITEMS.SLUG'
                   )
                 }}</label>
-                <span class="text-block-title">{{ portal.slug }}</span>
+                <span class="text-block-title">
+                  {{ portal.slug }}
+                </span>
               </div>
               <div class="configuration-item">
                 <label>{{
@@ -114,7 +122,9 @@
                     'HELP_CENTER.PORTAL.PORTAL_SETTINGS.LIST_ITEM.PORTAL_CONFIG.ITEMS.TITLE'
                   )
                 }}</label>
-                <span class="text-block-title">{{ portal.page_title }}</span>
+                <span class="text-block-title">
+                  {{ portal.page_title }}
+                </span>
               </div>
             </div>
             <div class="configuration-items">
@@ -137,7 +147,9 @@
                     'HELP_CENTER.PORTAL.PORTAL_SETTINGS.LIST_ITEM.PORTAL_CONFIG.ITEMS.SUB_TEXT'
                   )
                 }}</label>
-                <span class="text-block-title">{{ portal.header_text }}</span>
+                <span class="text-block-title">
+                  {{ portal.header_text }}
+                </span>
               </div>
             </div>
           </div>
@@ -159,6 +171,15 @@
         </div>
       </div>
     </div>
+    <woot-delete-modal
+      :show.sync="showDeleteConfirmationPopup"
+      :on-close="closeDeletePopup"
+      :on-confirm="onClickDeletePortal"
+      :title="$t('HELP_CENTER.PORTAL.PORTAL_SETTINGS.DELETE_PORTAL.TITLE')"
+      :message="deleteMessage"
+      :confirm-text="deleteConfirmText"
+      :reject-text="deleteRejectText"
+    />
   </div>
 </template>
 
@@ -166,12 +187,14 @@
 import thumbnail from 'dashboard/components/widgets/Thumbnail';
 import Label from 'dashboard/components/ui/Label';
 import LocaleItemTable from './PortalListItemTable';
+import alertMixin from 'shared/mixins/alertMixin';
 export default {
   components: {
     thumbnail,
     Label,
     LocaleItemTable,
   },
+  mixins: [alertMixin],
   props: {
     portal: {
       type: Object,
@@ -183,6 +206,13 @@ export default {
       values: ['archived', 'draft', 'published'],
     },
   },
+  data() {
+    return {
+      showDeleteConfirmationPopup: false,
+      alertMessage: '',
+      selectedPortalForDelete: {},
+    };
+  },
   computed: {
     labelColor() {
       switch (this.status) {
@@ -191,6 +221,22 @@ export default {
         default:
           return 'success';
       }
+    },
+    // Delete portal modal
+    deleteConfirmText() {
+      return `${this.$t(
+        'HELP_CENTER.PORTAL.PORTAL_SETTINGS.DELETE_PORTAL.YES'
+      )} ${this.selectedPortalForDelete.name}`;
+    },
+    deleteRejectText() {
+      return `${this.$t(
+        'HELP_CENTER.PORTAL.PORTAL_SETTINGS.DELETE_PORTAL.NO'
+      )} ${this.selectedPortalForDelete.name}`;
+    },
+    deleteMessage() {
+      return `${this.$t(
+        'HELP_CENTER.PORTAL.PORTAL_SETTINGS.DELETE_PORTAL.MESSAGE'
+      )} ${this.selectedPortalForDelete.name} ?`;
     },
 
     locales() {
@@ -207,8 +253,34 @@ export default {
     openSettings() {
       this.$emit('open');
     },
-    onClickDelete(portal) {
-      this.$emit('delete', portal);
+    onClickOpenDeleteModal(portal) {
+      this.selectedPortalForDelete = portal;
+      this.showDeleteConfirmationPopup = true;
+    },
+    closeDeletePopup() {
+      this.showDeleteConfirmationPopup = false;
+    },
+    async onClickDeletePortal() {
+      const { id, slug } = this.selectedPortalForDelete;
+      try {
+        await this.$store.dispatch('portals/delete', {
+          id: id,
+          slug: slug,
+        });
+        this.selectedPortalForDelete = {};
+        this.closeDeletePopup();
+        this.alertMessage = this.$t(
+          'HELP_CENTER.PORTAL.PORTAL_SETTINGS.DELETE_PORTAL.API.DELETE_SUCCESS'
+        );
+      } catch (error) {
+        this.alertMessage =
+          error?.message ||
+          this.$t(
+            'HELP_CENTER.PORTAL.PORTAL_SETTINGS.DELETE_PORTAL.API.DELETE_ERROR'
+          );
+      } finally {
+        this.showAlert(this.alertMessage);
+      }
     },
     swapLocale() {
       this.$emit('swap');
